mbedtls: add a new Kconfig file for PSA_WANT logic
The already existing "Kconfig.psa" maps Mbed TLS's PSA_WANT_xxx
symbols to Kconfigs that can be used in Zephyr to select which
PSA crypto API feature should be enabled in the build. In order
to ease maintainability, "Kconfig.psa" is automatically generated
so it should not be edited manually to add logic between
Kconfigs symbols defined there. As a consequence a new Kconfig file
is introduced in this commit, named "Kconfig.psa.logic", to
address this limitation. This new Kconfig file does not add new
public symbols (only hidden ones, if needed) and it simply adds
logic between PSA_WANT ones.

This commit also renames "Kconfig.psa" as "Kconfig.psa.auto" to
put it at the same "naming level" as the newly created file and,
at the same time, emphasize that it is an automatically generated
file.

Signed-off-by: Valerio Setti <vsetti@baylibre.com>