mbedtls: add Kconfig to select the number of key slot in PSA core
Adding new CONFIG_MBEDTLS_PSA_KEY_SLOT_COUNT to select the number
of key slots in PSA core. The default value is 32 which is the
same value used in Mbed TLS by default. Be aware that key slots
consume RAM memory even if unused, so the proper value should be
a compromise between the number of slots required by the application
and the available RAM in the system.

Signed-off-by: Valerio Setti <vsetti@baylibre.com>